+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### App Center
 
 * **[Fix]** Fix "Collection was mutated while being enumerated" exception in MSACChannelGroupDefault.
+* **[Fix]** Fix crash channel:didPrepareLog in MSACChannelGroupDefault
 
 ### App Center Distribute